add a decode_URI that will help to decode an encoded request path && upgate about code struct

+++ config_parsing/includes/ServerBlock.hpp
@@ -10,11 +10,11 @@
 struct ServerBlock
 {
     size_t listen;
-    size_t set_timeout;
+    int set_timeout;
     std::string root;
     std::string host;
     std::string server_name;
-    size_t client_max_body_size;
+    int client_max_body_size;
     std::set<std::string> index;
     std::map<int, std::string> error_page;
     std::deque<LocationBlock> locations;
